Software License Changes
The Creative Commons Attribution-NonCommercial 4.0 International
("CC BY-NC 4.0") license originally used with the Kaleidoscope software
and source code is not valid for software licensing. Since the original
license is not valid for software licensing, the copyright holder is now
upgrading the license from its original version.

As of August 28, 2016, the Kaleidoscope software will be licensed under
the GNU General Public License v3 ("GPLv3").

All previously distributed copies and modifications of this software are
hereby licensed under the GPLv3 license retroactively.